+++ vite.config.ts
@@ -21,5 +21,6 @@ export default defineConfig({
   build: {
     outDir: 'dist',
     assetsDir: 'assets',
+    chunkSizeWarningLimit: 700,
   },
 })